Some int -> NSInteger and float -> CGFloat transitions.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/trunk@40138 72102866-910b-0410-8b05-ffd578937521

+++ Headers/AppKit/NSTableHeaderView.h
@@ -39,7 +39,7 @@
 @interface NSTableHeaderView : NSView
 {
   NSTableView *_tableView;
-  int _resizedColumn;
+  NSInteger _resizedColumn;
 }
 /*
  * Setting the table view
